@@ -48,28 +48,41 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
         }
     });
 
-    Some(undistract_me(module, &config, elapsed))
+    Some(undistract_me(module, &config, context, elapsed))
 }
 
 #[cfg(not(feature = "notify"))]
-fn undistract_me<'a, 'b>(
+fn undistract_me<'a>(
     module: Module<'a>,
-    _config: &'b CmdDurationConfig,
+    _config: &CmdDurationConfig,
+    _context: &'a Context,
     _elapsed: u128,
 ) -> Module<'a> {
     module
 }
 
 #[cfg(feature = "notify")]
-fn undistract_me<'a, 'b>(
+fn undistract_me<'a>(
     module: Module<'a>,
-    config: &'b CmdDurationConfig,
+    config: &CmdDurationConfig,
+    context: &'a Context,
     elapsed: u128,
 ) -> Module<'a> {
     use notify_rust::{Notification, Timeout};
     use nu_ansi_term::{unstyle, AnsiStrings};
 
     if config.show_notifications && config.min_time_to_notify as u128 <= elapsed {
+        if cfg!(target_os = "linux") {
+            let in_graphical_session = ["DISPLAY", "WAYLAND_DISPLAY", "MIR_SOCKET"]
+                .iter()
+                .find_map(|&var| context.get_env(var).filter(|val| !val.is_empty()))
+                .is_some();
+
+            if !in_graphical_session {
+                return module;
+            };
+        }
+
         let body = format!(
             "Command execution {}",
             unstyle(&AnsiStrings(&module.ansi_strings()))
