@@ -45,7 +45,7 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
                 "style" => {
                     let engines_version = get_engines_version(context);
                     let in_engines_range =
-                        check_engines_version(nodejs_version.deref().as_ref()?, engines_version);
+                        check_engines_version(nodejs_version.as_deref(), engines_version);
                     if in_engines_range {
                         Some(Ok(config.style))
                     } else {
@@ -92,13 +92,12 @@ fn get_engines_version(context: &Context) -> Option<String> {
     Some(raw_version.to_string())
 }
 
-fn check_engines_version(nodejs_version: &str, engines_version: Option<String>) -> bool {
-    if engines_version.is_none() {
+fn check_engines_version(nodejs_version: Option<&str>, engines_version: Option<String>) -> bool {
+    let (Some(nodejs_version), Some(engines_version)) = (nodejs_version, engines_version) else {
         return true;
-    }
-    let r = match VersionReq::parse(&engines_version.unwrap()) {
-        Ok(r) => r,
-        Err(_e) => return true,
+    };
+    let Ok(r) = VersionReq::parse(&engines_version) else {
+        return true;
     };
     let re = Regex::new(r"\d+\.\d+\.\d+").unwrap();
     let version = re
@@ -267,6 +266,18 @@ mod tests {
 
         let actual = ModuleRenderer::new("nodejs").path(dir.path()).collect();
         let expected = Some(format!("via {}", Color::Red.bold().paint(" v12.0.0 ")));
+        assert_eq!(expected, actual);
+        dir.close()
+    }
+    #[test]
+    fn no_node_installed() -> io::Result<()> {
+        let dir = tempfile::tempdir()?;
+        File::create(dir.path().join("index.js"))?.sync_all()?;
+        let actual = ModuleRenderer::new("nodejs")
+            .path(dir.path())
+            .cmd("node --version", None)
+            .collect();
+        let expected = Some(format!("via {}", Color::Green.bold().paint(" ")));
         assert_eq!(expected, actual);
         dir.close()
     }
