@@ -13,6 +13,7 @@ use std::path::Path;
 /// 2a) (not implemented on macOS) one of the supplementary groups of the current user is the
 /// directory group owner and whether it has write access
 /// 3) 'others' part of the access mask has the write access
+#[allow(clippy::useless_conversion)] // On some platforms it is not u32
 pub fn is_write_allowed(folder_path: &Path) -> Result<bool, String> {
     let meta =
         fs::metadata(folder_path).map_err(|e| format!("Unable to stat() directory: {e:?}"))?;
@@ -22,14 +23,15 @@ pub fn is_write_allowed(folder_path: &Path) -> Result<bool, String> {
     if euid.is_root() {
         return Ok(true);
     }
+
     if meta.uid() == euid.as_raw() {
-        Ok(perms & Mode::S_IWUSR.bits() as u32 != 0)
+        Ok(perms & u32::from(Mode::S_IWUSR.bits()) != 0)
     } else if (meta.gid() == Gid::effective().as_raw())
         || (get_supplementary_groups().contains(&meta.gid()))
     {
-        Ok(perms & Mode::S_IWGRP.bits() as u32 != 0)
+        Ok(perms & u32::from(Mode::S_IWGRP.bits()) != 0)
     } else {
-        Ok(perms & Mode::S_IWOTH.bits() as u32 != 0)
+        Ok(perms & u32::from(Mode::S_IWOTH.bits()) != 0)
     }
 }
 
